add secu

+++ app_aws/lambdas/stock/deleteStockLambda.ts
@@ -1,5 +1,6 @@
 import { DynamoDB } from '@aws-sdk/client-dynamodb';
 import { DynamoDBDocument } from '@aws-sdk/lib-dynamodb';
+import * as jwt from 'jsonwebtoken';
 
 const db = DynamoDBDocument.from(new DynamoDB());
 const TableName = process.env.TABLE; 
@@ -15,6 +16,18 @@ export const handler = async (event : any) => {
         const stockId = event.pathParameters.stockId;
         switch (event.requestContext.httpMethod) {
             case 'DELETE':
+                let token = event.headers.Authorization;
+
+                if (token.startsWith('Bearer ')) {
+                    token = token.slice(7, token.length);
+                }
+        
+                const claims = jwt.decode(token) as jwt.JwtPayload;
+        
+                if (!claims || !(claims['custom:role'] && (claims['custom:role'].includes('orga') || claims['custom:role'].includes('admin')))) {
+                    throw { statusCode: 403, message: 'Access denied.'};
+                }
+
                 body = await db.delete({ TableName, Key: { "stock-id": stockId } });
                 body = {"msg" : "stock deleted"}
                 break;
